moved signup & login logic to service
