--- src/app/modules/user/pages/user-password-reset/user-password-reset.page.html
@@ -1,13 +1,0 @@
-<ion-header [translucent]="true">
-  <ion-toolbar>
-    <ion-title>UserPasswordReset</ion-title>
-  </ion-toolbar>
-</ion-header>
-
-<ion-content [fullscreen]="true">
-  <ion-header collapse="condense">
-    <ion-toolbar>
-      <ion-title size="large">UserPasswordReset</ion-title>
-    </ion-toolbar>
-  </ion-header>
-</ion-content>
